gdt/requirements.txt
asgiref==3.8.1
Django==5.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-jsonapi==7.0.2
djangorestframework-simplejwt==5.3.1
inflection==0.5.1
PyJWT==2.9.0
python-dateutil==2.9.0.post0
six==1.16.0
sqlparse==0.5.1